Add files via upload
This is the page for displaying photos for a specified event. Not quite 100% complete yet, still needs the upload and download functionality and functionality for retrieving the event information and photo URLs from the backend, but it does allow you to view a gallery of photos (dummy data) and click on a photo to see it in fullscreen.